Remove `tensor_traits` and rename `expression_traits` to `traits`.
We already needed `expression_traits` for the `Tensor` template, and these
provide enough information to use where `tensor_traits` were used. Rather than
supporting the two different traits implemenations we just rely on the `Tensor`
template's `expression_traits`. The traits for the `Tensor` template don't
define an external index type, so tensors can't accidentally be used as
expressions.

The `expression_traits` were already namespaced inside of `ttl::expressions`,
which won't be imported by the user for any reason, so the `expression_` prefix
is redundant and leads to unnecessarily long lines.