Update a static assert during IndexMap get() evaluation.
The IndexMap get() operation is used at runtime to map a specific index from the
outer space to the inner space. It may actually have any numer of unused
dimensions due to the way we generate indices during evaluation, this its type
may not be equivalent to the IndexMap's OuterType, but should merely be a
superset.

This commit alos adds documentation to the IndexMap and TensorMap templates.