Add a transpose function to the library.
The transpose operator can be applied to any expression and just reverses the
index binding.

`transpose(A(i,j,k))==A(k,j,i);`

Transpose does not work on raw Tensors though we could make that extension. The
transpose operator doesn't actually _perform_ a transpose unless it is evaluated
during an assignment---it just introduces an expression that remaps indices.

This closes #7.