Add a delta expression.
This expression allows users to specify delta expressions directly in
statements, without needing a Delta "Tensor", and is half of #13.

Tensor<3,4,double> D = delta<4>(i,j,k);

The delta expression does not allocate storage but merely performs checks during
evaluation to see if the index is diagonal, i.e., if the integer elements in the
index are all equivalent. The intent is that TTL knows all loop indices and
bounds and thus the compiler should be able to evaluate this expression and
return 0 or 1 as necessary.