Add functionality to deal with undimensional expressions, and update the delta function with that capability.
As we add custom expression nodes that represent things like the delta function
or epsilon we will encounter situations where we do not need to know the
dimensionality of the underlying data. For instance, the delta function simply
tests the index and returns 1 when all index values are equal.

We introduce the special dimension of -1 for these cases, and update our type
trait inference to account for this. This allows natural expressions like
`delta(i,j)` to be used without regard for the dimension.

There are still cases where the dimension is necessary, for instance whenever we
need to apply or contract an expression. We add asserts to make sure that
dimensions are available in that case, and any expression that can be part of a
raw contraction must provide the cabability to explicitly specify the
dimensions.

As an example the delta function can be contracted like
`delta<3>(i,j)*delta<3>(j,k)`.

This is related to #13.