We use the Bind expression node for more than one purpose.
Bind nodes capture references to Tensors and allow them to be used as
functions. The also capture other expressions and allow us to permute indices as
we decend through the expression tree.

This means that the captured value needs to be a reference for Tensors, but a
value for Expressions. This is important because expressions are mostly
anonymous temporaries and are destroyed at the end of a statement. If there is
an expression captured by an auto name and used later then none of the tree can
be a reference.

```
auto f = A(i,j).to(j,i);
auto s = f(0,1);
```

If the Bind node created by the `to(j,i)` operator captures the Bind node node
created by the `A(i,j)` by reference, then the use of `f` contains a dangling
reference. If it captures the child by value then it is fine as it is copied as
needed. On the other hand, if `A(i,j)` captures the Tensor `A` by value then it
is making copies of potentially large structures and can't be uses as a
left-hand-side.

This patch uses a type trait to figure out which way to capture the child.