Update the library solve API to report failures.
This exposes two different APIs for getting errors from ttl during linear sytem
solving. The first is the `x=solve(A,b);` form, where we will throw the lapack
error code if there is one. The second is `e=solve(A,b,x);` which treats `x` as
an output tensor and returns the error code.