Leverage `--expt-relaxed-constexpr` for CUDA HOST/DEVICE code.
A lot of TTL is designed with inline `constexpr` annotations. These functions
are always supposed to be both `__host__` and `__device__`. Using the above flag
tells `nvcc` that all `constexpr` functions are implicitly both `__host__` and
`__device__`.

This improves the readability, maintainability, and portability of the TTL
headers without introducting excess cruft.