GCC doesn't allow template integers to be computed during partial specialization.
This is apparently due to standard 14.5.4/9.

``` A partially specialized non-type argument expression shall not involve a
template parameter of the partial specialization except when the argument
expression is a simple identifier.  ```

We were failing to compile code that looks like.

```
template <int n, int N>
struct impl { ... };

template <int n>
struct impl<n, n + 2> { ... };
```

This commit changes the `forall` and `contract` functions to work with
`std::integral_types` instead of `ints` directly. It's more verbose, but leads
to the same result.