Reduce the flexibility of the TensorImpl.
There were a lot of constructors and assignment operators that were designed to
allow flexibility in terms of the form with which constructors are constructed
and copied. The idea behind these was to make a Tensor look as much like an
array as possible.

Most of these options were never used, and added a burden to maintenance and
understanding of the Tensor API. Removing them simplifies things. If we find
compelling cases where we would like to support the functionality that they
provided, we can add them back in.