Simplify TTL's operator implementation.
This reduces all of the previously varied classes of operators to binary, unary,
and products. We can do this because we upgraded the underlying types to use
lambda functions, rather than just using std function objects. These lambda
functions can capture state, so things that used to be binary scalar operators
can be encoded as unary ops with a captured scalar.